passage norme

+++ src/ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: pdamoune <pdamoune@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/08 13:46:21 by pdamoune          #+#    #+#             */
-/*   Updated: 2017/05/06 14:39:38 by pdamoune         ###   ########.fr       */
+/*   Updated: 2017/06/08 19:21:31 by pdamoune         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,5 +25,5 @@ void	ft_bzero(void *s, size_t n)
 	str = (char *)s;
 	i = 0;
 	while (i < n)
-		str[i++] = (__uint128_t)0;
+		str[i++] = 0;
 }
